+++ src/Client/ClientCore/Entities/ClientEntityBuilder.cs
@@ -40,10 +40,11 @@ public sealed class ClientEntityBuilder : AbstractEntityBuilder
         OrientationComponentCollection orientations,
         AdminTagCollection admins,
         BlockPositionComponentCollection blockPositions,
+        CastBlockShadowsTagCollection castBlockShadows,
         EntityTable entityTable)
         : base(entityId, isLoad, entityManager, kinematics, materials,
             materialModifiers, aboveBlocks, playerCharacterTags, names, parents, drawables, animatedSprites,
-            orientations, admins, blockPositions, entityTable)
+            orientations, admins, blockPositions, castBlockShadows, entityTable)
     {
     }