+++ src/Server/Persistence/State/Trackers/StateTrackerInstaller.cs
@@ -59,7 +59,7 @@ public sealed class StateTrackerInstaller : IWindsorInstaller
             .LifestyleSingleton());
 
         container.Register(Component.For<AdminStateTracker>().LifestyleSingleton());
-
         container.Register(Component.For<TemplateStateTracker>().LifestyleSingleton());
+        container.Register(Component.For<CastBlockShadowsStateTracker>().LifestyleSingleton());
     }
 }